refactor: localstorage 관련 로직 전부 utils에 이관
