+++ src/apis/baskets.js
@@ -1,12 +1,12 @@
 import request from './utils/request';
 import apis from './index';
 import auth from '../utils/auth';
-import localstorage, { STORAGE_KEYS } from '../utils/localstorage';
+import localstorage from '../utils/localstorage';
 
 const baskets = {
   async getAll() {
     const isGuest = !auth.isLoggedIn();
-    const baskets = isGuest ? localstorage.get(STORAGE_KEYS.BASKETS) || [] : await request.get('/baskets');
+    const baskets = isGuest ? localstorage.getBaskets() : await request.get('/baskets');
     const getProductPromise = baskets.map(({ product_id }) => apis.products.get(product_id));
     const products = await Promise.all(getProductPromise);
 
@@ -18,34 +18,19 @@ const baskets = {
   },
 
   async update(id, amount) {
-    if (auth.isLoggedIn()) {
-      return request.put(`/baskets/${id}`, { amount });
+    const isGuest = !auth.isLoggedIn();
+    if (isGuest) {
+      return localstorage.updateBasket(id, amount);
     }
-    const baskets = localstorage.get(STORAGE_KEYS.BASKETS) || [];
-    const targetIndex = baskets.findIndex((basket) => basket.id === id);
-    const changedBasket = {
-      ...baskets[targetIndex],
-      amount,
-    };
-    const changed = [
-      ...baskets.slice(0, targetIndex),
-      changedBasket,
-      ...baskets.slice(targetIndex + 1, baskets.length),
-    ];
-    localstorage.set(STORAGE_KEYS.BASKETS, changed);
+    return request.put(`/baskets/${id}`, { amount });
   },
 
   async remove(id) {
-    if (auth.isLoggedIn()) {
-      return request.delete(`/baskets/${id}`);
+    const isGuest = !auth.isLoggedIn();
+    if (isGuest) {
+      return localstorage.removeBasket();
     }
-    const baskets = localstorage.get(STORAGE_KEYS.BASKETS) || [];
-    if (baskets.length === 1) {
-      localstorage.set(STORAGE_KEYS.BASKETS, []);
-      return;
-    }
-    const targetIndex = baskets.findIndex((basket) => basket.id === id);
-    localstorage.set(STORAGE_KEYS.BASKETS, baskets.splice(targetIndex, 1));
+    return request.delete(`/baskets/${id}`);
   },
 };
